requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.3
asyncpg==0.26.0
attrs==23.1.0
bcrypt==4.0.1
blis==0.7.10
caio==0.9.12
catalogue==2.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
confection==0.1.1
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
cymem==2.0.7
databases==0.6.1
DAWG-Python==0.7.2
docopt==0.6.2
ecdsa==0.18.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
fastapi==0.95.2
fonttools==4.42.1
frozenlist==1.4.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
murmurhash==1.0.9
mypy==1.3.0
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.25.2
orjson==3.9.0
packaging==23.1
pandas==2.0.3
passlib==1.7.4
pathy==0.10.2
Pillow==10.0.0
plotly==5.16.1
pluggy==1.0.0
preshed==3.0.8
psycopg2-binary==2.9.3
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.0.9
pytest==7.3.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
regex==2023.8.8
requests==2.31.0
rsa==4.9
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.6.0/ru_core_news_sm-3.6.0-py3-none-any.whl#sha256=9d11147ab9380106f61bd0348443d714008be174a675d85de44c80dee4a74307
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
srsly==2.4.7
starlette==0.27.0
tenacity==8.2.3
tensorflow==2.14.0
tensorflow-addons==0.21.0
thinc==8.1.12
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
types-aiofiles==23.1.0.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.22.0
uvloop==0.17.0
wasabi==1.1.2
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
